requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
pydantic==2.5.0
pandas==2.1.4
openai==1.10.0
python-dotenv==1.0.0
python-dateutil==2.8.2
geojson==3.1.0